Add a patch (00_create-libdir) that fixes the upstream build system to make sure the LIB_DIR is created or bail out.
Update the 01_fix-makefile patch to only install libGL/libOSMesa if they
exist. Don't make installation dependent on which drivers are built.

Build optimized versions of the swx11 libraries in addition to those that
should work on the least-capable supported processor for the given
architecture.

Have most configurations include debian-default instead of debian-common so
that INSTALL_DIR is correctly defined. debian-common can probably be dropped
anyway.