No longer install the default build of libOSMesa in the libgl1-mesa-swx11 package, but in the libosmesa6 package.
Modify the 01_fix-makefile patch to not build libOSMesa on stand-alone builds,
so that it is only created when building OSMesa explicitly.

Build libOSMesa with 8 bits per color channel (the default) in dynamic and
static versions separately from the software rasterization libGL with the
software rasterization code linked in statically so that libOSMesa can run in
parallel with any libGL. This should fix the problem Yann Dirson mentioned a
while ago.

Use the install-source.sh script to install the Mesa sources needed for the
mesa-swx11-source package instead of doing it in the debian/rules file itself.