Avoid installing nouveau_vieux_dri.so into two packages
Moving nouveau_vieux_dri.so to build/dri/lib/gallium/ in the build
target has the disadvantage that the upstream Makefile will detect the
missing file and bring it back later when we run "make install".  So
let's make the move only when we install the files into the packages.